Revert "chore: add integ test that validates full proxy traversal (#33092)"
This reverts commit 9e220c85113602d3ed571a030a40b005c6b17590.
