+++ packages/@aws-cdk-testing/cli-integ/lib/package-sources/repo-source.ts
@@ -7,10 +7,9 @@ import { shell, addToShellPath } from '../shell';
 
 export class RepoPackageSourceSetup implements IPackageSourceSetup {
   readonly name = 'repo';
-  readonly description: string;
+  readonly description = `repo(${this.repoRoot})`;
 
   constructor(private readonly repoRoot: string) {
-    this.description = `repo(${this.repoRoot})`;
   }
 
   public async prepare(): Promise<void> {
