+++ packages/@aws-cdk-testing/cli-integ/tsconfig.json
@@ -2,7 +2,7 @@
     "compilerOptions": {
         "target": "ES2020",
         "module": "commonjs",
-        "lib": ["ES2020", "dom"],
+        "lib": ["es2019", "es2020", "dom"],
         "strict": true,
         "alwaysStrict": true,
         "declaration": true,
